Support Endless Staircase and Not Saving Mesh (#36)
Added support to _NOT_ save mesh to .tscn file which allows for
generating arbitrarily large meshes. The mesh is still added to the
editor scene for viewing and testing, but it will not be saved to the
.tscn file.

Refactored how the meshes would be saved such that the `MeshInstance3D`
node is now a child of the staircase instead of the staircase node
itself.