Web Build Patch Attempt 2 (#44)
Updated demo-project to godot version 4.4

Downgraded godot-cpp to version 4.1 for greater compatibility
* Also added regression 

Fixed bug in gdextension build workflow that was creating duplicate DLLs
in the addons folder that weren't cleaned up properly. These duplicate
DLLs were leading to issues during github actions related to
GDExtensions. GDExtensions was trying to duplicate the create operations
if there were duplicate DLLs for the same platform. I cleaned up the
collect extensions workflow to attempt to resolve this issue. Also
attempted to remove the symlink as part of debugging, but this didn't
seem to have any affect, will leave the symlink for now.